requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anthropic==0.77.1
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
bce-python-sdk==0.9.57
beautifulsoup4==4.14.3
bert-score==0.3.13
blessed==1.25.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cupy-cuda11x==13.6.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
decord==0.6.0
diffusers
distro==1.9.0
docstring_parser==0.17.0
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
et_xmlfile==2.0.0
fastrlock==0.8.3
filelock==3.20.0
flash-attn
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.59.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gpustat==1.1.1
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.81
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.5.1
lap==0.5.12
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lpips==0.1.4
lxml==6.0.2
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.9.3
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
natsort==8.4.0
networkx==3.6.1
numba==0.64.0
numpy==2.3.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.590.44
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai==2.15.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pillow==12.0.0
platformdirs==4.9.4
playwright==1.57.0
plotly==6.5.0
pooch==1.9.0
premailer==3.10.0
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.0
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyecharts==2.0.9
pyee==13.0.0
pyparsing==3.3.1
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.14
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.2
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
SQLAlchemy==2.0.45
sympy==1.13.1
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchdiffeq==0.2.5
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.57.3
triton==3.2.0
typer-slim==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
websockets==15.0.1
wordcloud==1.9.5
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
